[PLT-XXX] Reduce amount of flushes in dispatcher
This prevents looping over flushes in the dispatcher in the case that we've already successfully flushed, but some datagrams entered the buffer whilst we were flushing.

We want to do this in order to reduce the amount of flushes and thus keep our batch sizes larger to promote client-side aggregation and reduce metric volume on the server.