[PLT-XXX] Improve tag writing
This improves tag writing such that written tags can be re-parsed with the existing tag parsing logic. There were two cases where we were able to write tags that could not be parsed:
1. When the tag had a single colon
2. When the tag contained a space

This will reduce the amount of unparseable metrics we're writing, which will reduce the amount of metrics we're currently losing due to parsing failures.