[PLT-913] Swap Net:HTTP with HTTPX
This swaps the http library from Net:HTTP to HTTPX for prometheus sending.

We are doing this because Net:HTTP spawns a thread to track connection timeouts, which is not possible to due when Rails is shutting down. This was causing us to report errors at every shutdown as well as to miss sending the last batch of metrics before a shutdown. HTTPX works differently under the hood and is thus not prone to these errors.