[PLT-913] Fix error case
This shifts up the `raise_for_status` call on the response so that we do not need to be aware of the response class.

Fixes a `The dispatcher thread encountered an error NoMethodError: undefined method `status' for an instance of HTTPX::ErrorResponse` exception that was happening because we were always checking the response `status` regardless of the response class.